requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
sqlalchemy==2.0.32
reportlab==4.2.2
python-docx==1.1.2
google-generativeai==0.7.2
python-dotenv==1.0.1
